Extract shared punctuation rule helpers
