+++ src/rules/support/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod flow_collection;
 pub(crate) mod line_syntax;
 pub(crate) mod mapping_key_walker;
+pub(crate) mod punctuation;
 pub(crate) mod span_utils;
